data-shift-monitor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
# pydantic==2.5.0
google-cloud-logging==3.8.0
google-auth==2.23.4
prometheus-client==0.19.0

sentence-transformers==2.6.1
transformers==4.35.2
unbabel_comet==2.2.1
pandas==2.2.2
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
scipy==1.14.1
huggingface_hub==0.30.2
# numpy==1.26.4
protobuf==4.25.6
tqdm==4.66.5
hf_xet